+++ src/openassetio-core/include/openassetio/BatchElementError.hpp
@@ -42,7 +42,22 @@ class BatchElementError final {
   /// Possible classes of error.
   enum class ErrorCode {
     /// Fallback for uncommon errors.
-    kUnknown = OPENASSETIO_BatchErrorCode_kUnknown
+    kUnknown = OPENASSETIO_BatchErrorCode_kUnknown,
+    /**
+     * Error code used during @ref resolve "entity resolution" when the
+     * reference itself is valid, but it is not possible to retrieve
+     * data for the referenced @ref entity.
+     *
+     * This could be because it does not exist, or a read-only entity is
+     * being resolved for write. This code should not be used if an
+     * entity does not have a requested trait - simply do not set that
+     * trait in the resulting data.
+     *
+     * This code is also used during version finalisation and any other
+     * entity-based operations on a valid @ref entity_reference that
+     * fail for some reason.
+     */
+    kEntityResolutionError = OPENASSETIO_BatchErrorCode_kEntityResolutionError,
   };
 
   /// Error code indicating the class of error.