+++ tests/python/openassetio/test_batchelementerror.py
@@ -29,6 +29,7 @@ from openassetio import BatchElementError
 class Test_BatchElementError_ErrorCode:
     def test_code_values(self):
         assert int(BatchElementError.ErrorCode.kUnknown) == 128
+        assert int(BatchElementError.ErrorCode.kEntityResolutionError) == 129
 
 
 class Test_BatchElementError_inheritance: